SMP
find and initialize cpus. the test boots up all found cpus (controllable with
CPUS env variable) each of which print a hello message, spin for a moment, and
then fault to make sure the IDT/GDT are properly setup.

there is no synchronization yet. next step is to make the runtime/kernel
thread-safe.

+++ biscuit/syscall.go
@@ -116,7 +116,7 @@ func elf_sload(p *proc_t, hdr *elf_phdr) {
 			if rsz - i < len {
 				len = rsz - i
 			}
-			runtime.Memmove(dst, src, uint(len))
+			runtime.Memmove(dst, src, len)
 		}
 		p.page_insert(hdr.vaddr + i, pg, p_pg, perms, true)
 	}
